requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.2
dataclasses==0.6
deep-translator==1.11.4
distlib==0.3.8
Django==5.1.2
django-cors-headers==4.4.0
# Editable install with no version control (django-debug-toolbar==4.4.6)
-e /Users/macbook/django/mysite/src/django-debug-toolbar
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
djongo==1.2.31
dnspython==2.7.0
filelock==3.15.4
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rfc3986==1.5.0
setuptools==72.2.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.1
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.26.3
